refactor logic for child links, move links to separate file
