+++ client/app/containers/StyleGuide/StyleGuideIndex.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { componentLinks } from './ComponentLinks';
 import StyleGuideModal from './StyleGuideModal';
 import StyleGuideTabs from './StyleGuideTabs';
 import StyleGuideProgressBar from './StyleGuideProgressBar';
@@ -10,7 +11,6 @@ import StyleGuidePlaceholder from './StyleGuidePlaceholder';
 import StyleGuideMessages from './StyleGuideMessages';
 import StickyNav from '../../components/StickyNav';
 import NavLink from '../../components/NavLink';
-import ChildNavLink from '../../components/ChildNavLink';
 import StyleGuideSearch from './StyleGuideSearch';
 import StyleGuideSearchableDropdown from './StyleGuideSearchableDropdown';
 import StyleGuideLogos from './StyleGuideLogos';
@@ -23,263 +23,6 @@ import StyleGuideFormFields from './StyleGuideFormFields';
 import StyleGuideAccordions from './StyleGuideAccordions';
 import StyleGuideAlerts from './StyleGuideAlerts';
 
-export const componentLinks = [
-  {
-    anchor: '/styleguide#',
-    name: 'Introduction'
-  },
-  {
-    anchor: '#branding',
-    name: 'Branding',
-    subnav: [{
-      anchor: '#logos',
-      name: 'Logos'
-    }]
-  },
-  {
-    anchor: '#typography',
-    name: 'Typography',
-    subnav: [{
-      anchor: '#text_styles',
-      name: 'Text Styles'
-    },
-    {
-      anchor: '#font_family',
-      name: 'Font Family'
-    },
-    {
-      anchor: '#text_accessibility',
-      name: 'Text Accessibility'
-    }]
-  },
-  {
-    anchor: '#colors',
-    name: 'Colors',
-    subnav: [{
-      anchor: '#platte',
-      name: 'Platte'
-    },
-    {
-      anchor: '#logos',
-      name: 'logos'
-    }]
-  },
-  {
-    anchor: '#buttons',
-    name: 'Buttons',
-    subnav: [{
-      anchor: '#primary',
-      name: 'primary'
-    },
-    {
-      anchor: '#secondary',
-      name: 'secondary'
-    },
-    {
-      anchor: '#Disabled',
-      name: 'Disabled'
-    },
-    {
-      anchor: '#link_buttons',
-      name: 'Link Buttons'
-    },
-    {
-      anchor: '#toggle_buttons',
-      name: 'Toggle Buttons'
-    }]
-  },
-  {
-    anchor: '#search',
-    name: 'Search'
-  },
-  {
-    anchor: '#dropdowns',
-    name: 'Dropdown Menus',
-    subnav: [{
-      anchor: '#single_selection',
-      name: 'Search Single Selection'
-    },
-    {
-      anchor: '#multiple_selection',
-      name: 'Create and Search Multiple Selection'
-    }]
-  },
-  {
-    anchor: '#checkboxes',
-    name: 'Checkboxes',
-    subnav: [{
-      anchor: '#single',
-      name: 'Single'
-    },
-    {
-      anchor: '#vertical',
-      name: 'Vertical'
-    },
-    {
-      anchor: '#horizontal',
-      name: 'Horizontal'
-    },
-    {
-      anchor: '#required',
-      name: 'Required'
-    },
-    {
-      anchor: '#acknowledgements',
-      name: 'Acknowledgements'
-    }]
-  },
-  {
-    anchor: '#radios',
-    name: 'Radio Buttons',
-    subnav: [{
-      anchor: '#vertical',
-      name: 'Vertical'
-    },
-    {
-      anchor: '#horizontal',
-      name: 'Horizontal'
-    }]
-  },
-  {
-    anchor: '#date_input',
-    name: 'Date Input'
-  },
-  {
-    anchor: '#tables',
-    name: 'Tables',
-    subnav: [{
-      anchor: '#queues',
-      name: 'Queues'
-    }]
-  },
-  {
-    anchor: '#tabs',
-    name: 'Tabs',
-    subnav: [{
-      anchor: '#without_icons',
-      name: 'Without Icons'
-    },
-    {
-      anchor: '#With_icons',
-      name: 'With Icons'
-    }]
-  },
-  {
-    anchor: '#accordions',
-    name: 'Accordions',
-    subnav: [{
-      anchor: '#border',
-      name: 'Border'
-    },
-    {
-      anchor: '#borderless',
-      name: 'Borderless'
-    }]
-  },
-  {
-    anchor: '#form_fields',
-    name: 'Form Fields',
-    subnav: [{
-      anchor: '#text_input',
-      name: 'Text input'
-    },
-    {
-      anchor: '#text_area',
-      name: 'Text area'
-    },
-    {
-      anchor: '#character_limit',
-      name: 'Character limit'
-    },
-    {
-      anchor: '#inline_form',
-      name: 'Inline Form'
-    }]
-  },
-  {
-    anchor: '#loaders',
-    name: 'Loaders',
-    subnav: [{
-      anchor: '#small_loader',
-      name: 'Small loader'
-    },
-    {
-      anchor: '#loading_buttons',
-      name: 'Loading buttons'
-    }]
-  },
-  {
-    anchor: '#alerts',
-    name: 'Alerts',
-    subnav: [{
-      anchor: '#alerts_lite',
-      name: 'Alerts lite'
-    }]
-  },
-  {
-    anchor: '#layout',
-    name: 'Layout',
-    subnav: [{
-      anchor: '#navigation_bar',
-      name: 'Navigation Bar'
-    },
-    {
-      anchor: '#user_dropdown_menu',
-      name: 'User dropdown menu'
-    },
-    {
-      anchor: '#content_area',
-      name: 'Main content area'
-    },
-    {
-      anchor: '#app_canvas',
-      name: 'App canvas'
-    },
-    {
-      anchor: '#actions',
-      name: 'Actions'
-    },
-    {
-      anchor: '#horizontal_line',
-      name: 'Horizontal line'
-    },
-    {
-      anchor: '#footer',
-      name: 'Footer'
-    }]
-  },
-  {
-    anchor: '#messages',
-    name: 'Messages',
-    subnav: [{
-      anchor: '#success_messages',
-      name: 'Success'
-    },
-    {
-      anchor: '#status_messages',
-      name: 'Status'
-    },
-    {
-      anchor: '#alerts',
-      name: 'Alert'
-    }
-    ]
-  },
-
-  {
-    anchor: '#dashboard',
-    name: 'Dashboard'
-  },
-  {
-    anchor: '#modals',
-    name: 'Modals'
-  },
-  {
-    anchor: '#progress_bar',
-    name: 'Progress Bar'
-  }
-];
-
 export default class StyleGuideIndex extends React.Component {
 
   render() {
@@ -288,8 +31,7 @@ export default class StyleGuideIndex extends React.Component {
         {
           componentLinks.map((link, i) => (
             <li key={i}>
-              <NavLink anchor={link.anchor} name={link.name} />
-              { link.subnav && <ChildNavLink links={link.subnav} key={link.name} /> }
+              <NavLink {...link} />
             </li>
           ))
         }
@@ -366,8 +108,6 @@ export default class StyleGuideIndex extends React.Component {
         <StyleGuideModal />
         <div className="cf-help-divider"></div>
         <StyleGuideProgressBar />
-        <div className="cf-help-divider"></div>
-
       </div>
     </div>;
   }
